Refactor component to use the same template both for single and multiple select
