+++ addon/components/power-select/before-options.js
@@ -1,0 +1,7 @@
+import Ember from 'ember';
+import layout from '../../templates/components/power-select/before-options';
+
+export default Ember.Component.extend({
+  tagName: '',
+  layout
+});
