+++ addon/templates/components/power-select.hbs
@@ -14,6 +14,7 @@
     noMatchesMessage=noMatchesMessage
     searchMessage=searchMessage
     selectedComponent=selectedComponentOrDefault
+    beforeOptionsComponent=beforeOptionsComponentOrDefault
     optionsComponent=optionsComponent
     matcher=matcher
     searchField=searchField
