+++ app/components/power-select/before-options.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-power-select/components/power-select/before-options';
